modify: 회원 DTO 수정 및 서비스 로직 개선

+++ src/main/java/com/nhnacademy/byeol23backend/couponset/couponpolicy/domain/dto/CouponPolicyCreateRequest.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public record CouponPolicyCreateRequest(
         String policyName,
-        Long criterionPrice,
+        BigDecimal criterionPrice,
         Integer discountRate,
         BigDecimal discountLimit,
         BigDecimal discountAmount
